refactor(prompt): replace factory pattern with handler pattern
- Rename PromptFactory to PromptHandler for clearer semantics
- Move factory.rs to handler.rs to better reflect the pattern
- Update module references and documentation accordingly
- Consolidate TemplateSource from factory.rs into loader/mod.rs
