--- src/loader/factory.rs
@@ -1,40 +1,0 @@
-use std::path::PathBuf;
-
-/// Factory for creating and managing template loaders based on source type.
-#[derive(Debug)]
-pub enum TemplateSource {
-    /// Local filesystem template path
-    FileSystem(PathBuf),
-    /// Git repository URL (HTTPS or SSH)
-    Git(String),
-}
-
-impl std::fmt::Display for TemplateSource {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            TemplateSource::FileSystem(path) => {
-                write!(f, "local path: '{}'", path.display())
-            }
-            TemplateSource::Git(repo) => write!(f, "git repository: '{repo}'"),
-        }
-    }
-}
-
-impl TemplateSource {}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_template_source_display() {
-        let fs_source = TemplateSource::FileSystem(PathBuf::from("/path/to/template"));
-        assert_eq!(format!("{}", fs_source), "local path: '/path/to/template'");
-
-        let git_source = TemplateSource::Git("git@github.com:user/repo".to_string());
-        assert_eq!(
-            format!("{}", git_source),
-            "git repository: 'git@github.com:user/repo'"
-        );
-    }
-}
